requirements.txt
azure-core==1.30.2
azure-cosmos==4.7.0
blinker==1.8.2
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.12.0
Flask==3.0.3
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
MarkupSafe==2.1.5
pip-upgrade==0.0.6
platformdirs==3.2.0
pur==7.1.0
PyYAML==6.0.2
requests==2.28.2
six==1.16.0
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==1.26.15
virtualenv==20.22.0
virtualenv-clone==0.5.7
Werkzeug==3.0.3